Reworked The first 2 parts
I reworked my code to a grid. I started with the banner part because i was having trouble getting it right but that is fully done. Then i kinda forgot to instantly commit so i continued to the header, added a background and added buttons for navigaton also i  added the logo.